Add previous and next functions and wrapping to the ViewStack
If wrap is set to true, when the users call previous or next, the stack
will wrap. If we are showing element 0 and previous is called then we
show the last element. Inversely, if we are showing the last element
and next is called then we show the first element.